requirements.txt
certifi==2018.11.29
Django==2.1.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-i18n-model==0.0.9
django-js-asset==1.2.2
django-location-field==2.1.0
django-map-widgets==0.2.2
django-mathfilters==0.4.0
django-modeltranslation==0.13.1
django-pagemore==0.2.0
django-pagination-bootstrap==1.3
django-suit==0.2.25
django-translations==1.1.1
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.8.2
pytz==2018.5
six==1.12.0
tqdm==4.31.1
whitenoise==4.1.2
wincertstore==0.2
